Move common handlers to server package
