+++ server/handler.go
@@ -1,4 +1,4 @@
-package handler
+package server
 
 import (
 	"net/http"
